refactor useFeed
